Convert prettyprinter-ansi-terminal
