+++ prettyprinter-ansi-terminal/src/Data/Text/Prettyprint/Doc/Render/Terminal.hs
@@ -1,35 +1,10 @@
--- | Render 'SimpleDocStream' in a terminal.
+-- | This module is part of the old @prettyprinter-ansi-terminal@ API which is
+-- planned to be deprecated and eventually removed.
+--
+-- Use "Prettyprinter.Render.Terminal" instead.
 module Data.Text.Prettyprint.Doc.Render.Terminal (
-    -- * Styling
-    AnsiStyle,
-    Color(..),
-
-    -- ** Font color
-    color, colorDull,
-
-    -- ** Background color
-    bgColor, bgColorDull,
-
-    -- ** Font style
-    bold, italicized, underlined,
-
-    -- ** Internal markers
-    --
-    -- | These should only be used for writing adaptors to other libraries; for
-    -- the average use case, use 'bold', 'bgColorDull', etc.
-    Intensity(..),
-    Bold(..),
-    Underlined(..),
-    Italicized(..),
-
-    -- * Conversion to ANSI-infused 'Text'
-    renderLazy, renderStrict,
-
-    -- * Render directly to 'stdout'
-    renderIO,
-
-    -- ** Convenience functions
-    putDoc, hPutDoc,
+    module Prettyprinter.Render.Terminal
 ) where
 
-import Data.Text.Prettyprint.Doc.Render.Terminal.Internal
+import Prettyprinter.Render.Terminal
+
